requirements.txt
Django~=4.2.7
django-environ==0.9.0
django-redis
redis
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
django-filter==22.1
psycopg2-binary==2.9.5
django-mptt
